requirements.txt
Pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.5.0
easyocr>=1.6.0
numpy>=1.21.0
